Update the userdata search paths
Add a Lambda-specific search path to accommodate Lambdas that are not
using an environment variable to supply userdata.

I had assumed that `File.join(__dir__, 'userdata.yml')` would take care
of both the EC2 and Lambda use cases, but this turns out to not be the
case. I suspect that since the entry point/shim scripts that the Lambdas
use are `cd`-ing to the `/tmp` directory first before invoking the EZQ
Processor, that may be subtly changing the output we get back from the
`Kernel.__dir__` method. To address this, just add an explicit check for
`/var/task/userdata.yml`.